+++ src/components/App.js
@@ -1,5 +1,6 @@
 import React, { useReducer, useState } from "react"
 import "bootstrap/dist/css/bootstrap.min.css"
+import Event from "./Event"
 import reducer from "../reducers"
 
 const App = () => {
@@ -65,31 +66,9 @@ const App = () => {
           </tr>
         </thead>
         <tbody>
-          {state.map((event, index) => {
-            const id = event.id
-            const handleClickDeleteButton = () => {
-              dispatch({
-                type: "DELETE_EVENT",
-                id: id,
-              })
-            }
-            return (
-              <tr key={index}>
-                <td>{id}</td>
-                <td>{event.title}</td>
-                <td>{event.body}</td>
-                <td>
-                  <button
-                    type="button"
-                    className="btn btn-danger"
-                    onClick={handleClickDeleteButton}
-                  >
-                    削除
-                  </button>
-                </td>
-              </tr>
-            )
-          })}
+          {state.map((event, index) => (
+            <Event key={index} event={event} dispatch={dispatch} />
+          ))}
         </tbody>
       </table>
     </div>
