+++ src/components/Event.js
@@ -1,0 +1,29 @@
+import React from "react"
+
+const Event = ({ event, dispatch }) => {
+  const id = event.id
+  const handleClickDeleteButton = () => {
+    dispatch({
+      type: "DELETE_EVENT",
+      id: id,
+    })
+  }
+  return (
+    <tr>
+      <td>{id}</td>
+      <td>{event.title}</td>
+      <td>{event.body}</td>
+      <td>
+        <button
+          type="button"
+          className="btn btn-danger"
+          onClick={handleClickDeleteButton}
+        >
+          削除
+        </button>
+      </td>
+    </tr>
+  )
+}
+
+export default Event
